multi: clean up types
Having the types all at the top-level was starting to muddy things up
quite a bit. This is quite a big change, but it tries to put the types
as close to the module that they are related to. We can re-export these
at the top-level to simplify things.
