+++ src/api/index.ts
@@ -1,2 +1,10 @@
+import { ComposeService } from './composes/service';
+
 export { composes } from './composes';
 export { meta } from './meta';
+
+export const services = {
+  Compose: ComposeService,
+};
+
+export * from './composes/types';
