+++ src/app.ts
@@ -3,12 +3,14 @@ import { pinoLogger } from 'hono-pino';
 import { prettyJSON } from 'hono/pretty-json';
 
 import * as api from '@app/api';
-import { ComposeService } from '@app/api/composes/service';
+import { ComposeRequest, services } from '@app/api';
 import { API_ENDPOINT } from '@app/constants';
 import { notFound, onError } from '@app/errors';
 import { logger } from '@app/logger';
 import { createQueue } from '@app/queue';
-import { AppContext, ComposeRequest, Store, Worker } from '@app/types';
+import { Store } from '@app/store';
+import { AppContext } from '@app/types';
+import { Worker } from '@app/worker';
 
 export const createApp = (
   socket: string,
@@ -16,7 +18,7 @@ export const createApp = (
   worker: Worker<ComposeRequest>,
 ) => {
   const queue = createQueue(worker);
-  const composeService = new ComposeService(queue, store);
+  const composeService = new services.Compose(queue, store);
 
   const middleware = new Hono<AppContext>();
   middleware.use(prettyJSON());
