+++ src/constants.ts
@@ -3,3 +3,10 @@ export const SOCKET_PATH = '/run/decomposer-httpd.sock';
 
 // Default store
 export const STORE_PATH = '/var/lib/osbuild-composer/artifacts';
+
+export enum Status {
+  SUCCESS = 'success',
+  FAILURE = 'failure',
+  PENDING = 'pending',
+  BUILDING = 'building',
+}
