+++ src/queue/index.ts
@@ -1,4 +1,5 @@
-import { ComposeRequest, Worker } from '@app/types';
+import { ComposeRequest } from '@app/api/composes';
+import { Worker } from '@app/worker';
 
 import { JobQueue } from './queue';
