+++ src/store.ts
@@ -1,7 +1,21 @@
 import path from 'path';
 import pouchdb from 'pouchdb';
 
-import { ComposeDoc } from '@app/types';
+import { ComposeRequest } from '@app/api/composes';
+import { Status } from '@app/constants';
+
+export type ComposeDoc = {
+  _id: string;
+  _rev?: string;
+  created_at: string;
+  status: Status;
+  request?: ComposeRequest;
+};
+
+export type Store = {
+  path: string;
+  composes: PouchDB.Database<ComposeDoc>;
+};
 
 export const createStore = (store: string) => {
   const composesPath = path.join(store, 'composes');
